move proveRequest + rename + refactor getViableAccountCredentialsForStatement to be partially usable in age proof
