+++ packages/browser-wallet/src/shared/utils/proof-helpers.ts
@@ -1,5 +1,16 @@
-import { AtomicStatement, StatementTypes } from '@concordium/web-sdk';
+import {
+    AtomicStatement,
+    CommitmentInput,
+    CryptographicParameters,
+    StatementTypes,
+    Web3IdProofInput,
+    Web3IdProofRequest,
+} from '@concordium/web-sdk';
+import { popupMessageHandler } from '@popup/shared/message-handler';
 import { ConfirmedIdentity } from '@shared/storage/types';
+import { InternalMessageType } from '@concordium/browser-wallet-message-hub';
+import { stringify } from '@concordium/browser-wallet-api/src/util';
+import { BackgroundResponseStatus, ProofBackgroundResponse } from './types';
 
 export function canProveStatement(statement: AtomicStatement, identity: ConfirmedIdentity) {
     const attribute = identity.idObject.value.attributeList.chosenAttributes[statement.attributeTag];
@@ -16,4 +27,30 @@ export function canProveStatement(statement: AtomicStatement, identity: Confirme
         default:
             throw new Error(`Statement type of ${statement.type} is not supported`);
     }
+}
+
+/**
+ * Helper to prove a Web3 request and get a verifiablePresentation.
+ * This returns the presentation as a JSON string.
+ */
+export async function proveWeb3Request(
+    request: Web3IdProofRequest,
+    commitmentInputs: CommitmentInput[],
+    global: CryptographicParameters
+) {
+    const input: Web3IdProofInput = {
+        request,
+        commitmentInputs,
+        globalContext: global,
+    };
+
+    const result: ProofBackgroundResponse<string> = await popupMessageHandler.sendInternalMessage(
+        InternalMessageType.CreateWeb3IdProof,
+        stringify(input)
+    );
+
+    if (result.status !== BackgroundResponseStatus.Success) {
+        throw new Error(result.reason);
+    }
+    return result.proof;
 }
